# ⚡️ Performance optimizations